backend/aiml_bot/requirements.txt
Flask==2.3.2
python-aiml==0.9.3
flask-cors==3.0.10
